# backend/src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "forum API",
    "version": "0.0.1",
    "description": "quipu backoffice API용 Swagger"
  },
  "paths": {
...
    "/bo/member": {
      "get": {
        "summary": "동아리 신청 인원 조회",
        "description": "member 테이블 조회",
        "responses": {
          "200": {
            "description": "조회 성공",
            "content": {
              "application/json": {
                "example": {
                  "message": "조회 성공"
                }
              }
            }
          },
          "500": {
            "description": "서버 에러",
            "content": {
              "application/json": {
                "example": {
                  "message": "서버 에러"
                }
              }
            }
          }
        }
      }
    },
...
  "/bo/semina": {
    "post": {
      "summary": "세미나 정보 등록 및 파일 업로드",
      "description": "세미나 정보를 DB에 저장, 관련 파일을 Cloudflare R2에 업로드",
      "consumes": ["multipart/form-data"],
      "requestBody": {
        "content": {
          "multipart/form-data": {
            "schema": {
              "type": "object",
              "properties": {
                "speaker": {
                  "type": "string",
                  "example": "swagger테스터"
                },
                "topic": {
                  "type": "string",
                  "example": "swagger테스트"
                },
                "detail": {
                  "type": "string",
                  "example": "swagger로 테스트 중임"
                },
                "resources": {
                  "type": "string",
                  "example": "https://example.com/resources.pdf"
                },
                "presentation_date": {
                  "type": "string",
                  "format": "date",
                  "example": "2025-03-01"
                },
                "files": {
                  "type": "array",
                  "items": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        }
      },
      "responses": {
        "200": {
          "description": "데이터 저장 및 파일 업로드 성공",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "message": {
                    "type": "string",
                    "example": "데이터 저장 및 파일 업로드 성공"
                  },
                  "semina": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "integer",
                        "example": 1
                      },
                      "speaker": {
                        "type": "string",
                        "example": "swagger테스터"
                      },
                      "topic": {
                        "type": "string",
                        "example": "swagger테스트"
                      },
                      "presentation_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-03-01"
                      }
                    }
                  },
                  "files": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "example": "original_filename.pdf"
                    }
                  }
                }
              }
            }
          }
        },
        "400": {
          "description": "잘못된 요청",
          "content": {
            "application/json": {
              "example": {
                "message": "업로드할 파일이 없습니다."
              }
            }
          }
        },
        "500": {
          "description": "서버 오류",
          "content": {
            "application/json": {
              "example": {
                "message": "서버 오류"
              }
            }
          }
        }
      }
    }
  },
...
  },
...
}